move power bi assets to its own file
