+++ hooli_data_eng/definitions.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from dagster._core.definitions.asset_spec import replace_attributes
 from dagster import (
     AnchorBasedFilePathMapping,
     Definitions,
@@ -8,22 +7,11 @@ from dagster import (
     build_column_schema_change_checks,
     multiprocess_executor,
     with_source_code_references,
-    EnvVar,
-    AssetSpec,
-    AssetKey,
 )
 from dagster_cloud.metadata.source_code import link_code_references_to_git_if_cloud
-from dagster_powerbi import (
-    PowerBIServicePrincipal,
-    PowerBIToken,
-    load_powerbi_asset_specs,
-    DagsterPowerBITranslator,
-    build_semantic_model_refresh_asset_definition,
-)
-from dagster_powerbi.translator import PowerBIContentData
-
 
 from hooli_data_eng.assets import forecasting, raw_data, marketing, dbt_assets
+from hooli_data_eng.assets.powerbi_assets import powerbi_assets
 from hooli_data_eng.assets.dbt_assets import dbt_slim_ci_job
 from hooli_data_eng.assets.marketing import check_avg_orders
 from hooli_data_eng.assets.raw_data import check_users, raw_data_schema_checks
@@ -34,7 +22,6 @@ from hooli_data_eng.sensors import orders_sensor, dbt_code_version_sensor
 from hooli_data_eng.sensors.watch_s3 import watch_s3_sensor
 from hooli_data_eng.assets.marketing import avg_orders_freshness_check, min_order_freshness_check, min_order_freshness_check_sensor, check_avg_orders, avg_orders_freshness_check_schedule
 from hooli_data_eng.assets.dbt_assets import weekly_freshness_check, weekly_freshness_check_sensor
-from hooli_data_eng.powerbi_workspace import power_bi_workspace
 # ---------------------------------------------------
 # Assets
 
@@ -71,31 +58,6 @@ forecasting_assets = load_assets_from_package_module(
 
 marketing_assets = load_assets_from_package_module(marketing, group_name="MARKETING")
 
-class MyCustomPowerBITranslator(DagsterPowerBITranslator):
-    def get_report_spec(self, data: PowerBIContentData) -> AssetSpec:
-        return super().get_report_spec(data)._replace(group_name="FORECASTING")
-
-    def get_semantic_model_spec(self, data: PowerBIContentData) -> AssetSpec:       
-        spec = super().get_semantic_model_spec(data)
-        return replace_attributes(spec, group_name="FORECASTING", deps=[AssetKey(path=[dep.asset_key.path[1].upper(), dep.asset_key.path[2]]) for dep in spec.deps])
-
-    def get_dashboard_spec(self, data: PowerBIContentData) -> AssetSpec:
-        return super().get_dashboard_spec(data)._replace(group_name="FORECASTING")
-    
-    def get_data_source_spec(self, data: PowerBIContentData) -> AssetSpec:
-        return super().get_data_source_spec(data)._replace(group_name="FORECASTING")
-
-
-power_bi_specs = load_powerbi_asset_specs(
-    power_bi_workspace, dagster_powerbi_translator=MyCustomPowerBITranslator, use_workspace_scan=True
-)
-
-power_bi_assets = [
-    build_semantic_model_refresh_asset_definition(resource_key="power_bi", spec=spec)
-    if spec.tags.get("dagster-powerbi/asset_type") == "semantic_model"
-    else spec
-    for spec in power_bi_specs
-]
 
 # ---------------------------------------------------
 # Definitions
@@ -108,7 +70,7 @@ defs = Definitions(
         {"max_concurrent": 3}
     ),  
     assets=link_code_references_to_git_if_cloud(
-        with_source_code_references([*dbt_assets, *raw_data_assets, *forecasting_assets, *marketing_assets, *power_bi_assets]),
+        with_source_code_references([*dbt_assets, *raw_data_assets, *forecasting_assets, *marketing_assets, *powerbi_assets]),
         file_path_mapping=AnchorBasedFilePathMapping(
             local_file_anchor=Path(__file__),
             file_anchor_path_in_repository="hooli_data_eng/definitions.py"
